Handle IPNS names without dots
When checking if an ipfs name is blocked, we may get an
`/ipns/{dnslink_name}` where the `{dnslink_name}` has been
encoded according to

https://specs.ipfs.tech/http-gateways/subdomain-gateway/#host-request-header

That is, "." has been replaced with "-" and "-" has been replaced with "--".

Because of this, when the given dns name has no dots, we undo things are
replace "--" with "-" and "-" with ".".