Fix: doublehashes can use different hashes
Currently, we tried to parse double-hashes as CIDv0s, which implies that
SHA256 was used to hash the double-hash. Instead we should try to parse a
multihash directly, and support any hashing function for the double-hash.

The rule14 test has been changed to use a blake3 double-hash, which previously
broke at the parsing level.